+++ js/length-check.js
@@ -1,0 +1,3 @@
+function checkStringLength (string, length) {
+  return string.length <= length;
+}
